+++ src/model.js
@@ -10,13 +10,17 @@ const { validateRequestQuery } = require('./helpers/validate-request-query');
 
 const MAX_PAGE_SIZE = 100; // maximum number of results returned from portal per request
 const FIELDS_DEFINITION = require('./fields-definition');
-const DEFAULT_PORTAL_URL = 'http://www.arcgis.com/sharing/rest/search';
+const PORTAL_ENDPOINTS = {
+  prod: 'http://www.arcgis.com/sharing/rest/search',
+  qa: 'https://qaext.arcgis.com/sharing/rest/search',
+  dev: 'https://devext.arcgis.com/sharing/rest/search'
+};
 
 class ArcgisSearchModel {
   constructor(koop, options = {}) {
     this.log = koop.log;
     this.ttl = options.ttl || 0;
-    this.portalUrl = options.portalUrl || DEFAULT_PORTAL_URL;
+    this.portalUrl = PORTAL_ENDPOINTS[options.portalEnv] || PORTAL_ENDPOINTS['prod'];
     this.userAgent = options.userAgent && setUserAgentForPortalRequest(options.userAgent);
   }
   // Main getData method which is used to send data to Koop