+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.0.2] - 2023-03-31
 ### Added
 * Support for stringifield valid geometry JSON object in request query